mptcp: update scheduler API
This patchset proposes an updated API for mptcp schedulers to provide
more flexibility for future scheduling strategies. This does not change
the current behavior of get_subflow. It adds an optional hook to control
scheduling behavior after selecting a subflow in get_subflow.

The main idea is to provide a way to perform a scheduling per-MSS or per
chunk of data. To demonstrate this, the mptcp_bpf_rr scheduler has been
modified to send one MSS per subflow in an alternating fashion.

The other potential application of this API could be to define a
scheduler that will proactivelly reinject data on available subflows
based on some heuristics (eg. when one subflow RTT increases when moving
away from WiFi AP). It could also allow to limit the usage of a specific
subflow, eg., in the case of high BDP.

I am mainly looking for feedback to see whether this design make sense.
Matthieu already did a quick pass on it.

The API introduces a push() function (name TBD) that takes the subflow
and data chunk (sequence number and length) as arguments. The scheduler
can then control packet scheduling by:
- Setting data transmission limits (e.g., per-MSS on a subflow or per window).
- Setting flags associated with the data. Currently, only
  MPTCP_SCHED_FLAG_RESCHEDULE is defined, forcing the scheduler to call
  get_subflow after sending the chunk.

I am envisioning another flag MPTCP_SCHED_FLAG_REINJECT that will ensure
that chunk of data can be rescheduled directly on another subflow.

I've written some packetdrill tests to validate the update RR scheduler
the test are available on my fork:

https://github.com/gdetal/packetdrill/tree/mptcp-net-next/gtests/net/mptcp-bpf/round-robin

Open questions:
- Is the current API for setting flags and limits appropriate? What are
  eBPF best practices to pass back information to kernel ?
- should get_scheduler and this new API be merged somehow as to provide
  a single way of scheduling data ? Eg. could pass chunk and return
  flags, limit and subflow to send this data to.

Co-Developed-By: Matthieu Baerts <matttbe@kernel.org>
Signed-off-by: Gregory Detal <gregory.detal@gmail.com>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20240527-sched_per_packet-v1-0-09a41d405f7c@gmail.com



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240503-sched_per_packet-d6e4488e54e8",
    "prefixes": [
      "mptcp-next RFC"
    ],
    "history": {
      "v1": [
        "20240527-sched_per_packet-v1-0-09a41d405f7c@gmail.com"
      ]
    }
  }
}