mptcp: allow to force a direct reinjection
MPTCP_SCHED_FLAG_REINJECT is a flag that allows a scheduler to
force a data reinjection directly. It implies MPTCP_SCHED_FLAG_RESCHEDULE.

The flag must be used with with care as one could infinitelly reinject
the same data forever. The following change updates the red bpf scheduler
to provide a proper example.

Signed-off-by: Gregory Detal <gregory.detal@gmail.com>